+++ src/components/Fields/FieldDetails.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 
 import {
-  NullableLabel,
   PatternLabel,
   RecursiveLabel,
   TypeFormat,
@@ -9,6 +8,8 @@ import {
   TypePrefix,
   TypeTitle,
   ToggleButton,
+  FieldLabel,
+  ExampleValue,
 } from '../../common-elements/fields';
 import { serializeParameterValue } from '../../utils/openapi';
 import { ExternalDocumentation } from '../ExternalDocumentation/ExternalDocumentation';
@@ -23,6 +24,8 @@ import { Badge } from '../../common-elements/';
 
 import { l } from '../../services/Labels';
 import { OptionsContext } from '../OptionsProvider';
+import { FieldModel } from '../../services/models/Field';
+import styled from '../../styled-components';
 
 const MAX_PATTERN_LENGTH = 45;
 
@@ -42,22 +45,21 @@ export class FieldDetails extends React.PureComponent<FieldProps, { patternShown
   render() {
     const { showExamples, field, renderDiscriminatorSwitch } = this.props;
     const { patternShown } = this.state;
-    const { enumSkipQuotes, hideSchemaTitles } = this.context;
+    const { enumSkipQuotes, hideSchemaTitles, hideSchemaPattern } = this.context;
 
-    const { schema, description, example, deprecated } = field;
+    const { schema, description, example, deprecated, examples } = field;
 
     const rawDefault = !!enumSkipQuotes || field.in === 'header'; // having quotes around header field default values is confusing and inappropriate
 
-    let exampleField: JSX.Element | null = null;
+    let renderedExamples: JSX.Element | null = null;
 
-    if (showExamples && example !== undefined) {
-      const label = l('example') + ':';
-      if (field.in && (field.style || field.serializationMime)) {
-        // decode for better readability in examples: see https://github.com/Redocly/redoc/issues/1138
-        const serializedValue = decodeURIComponent(serializeParameterValue(field, example));
-        exampleField = <FieldDetail label={label} value={serializedValue} raw={true} />;
+    if (showExamples && (example !== undefined || examples !== undefined)) {
+      if (examples !== undefined) {
+        renderedExamples = <Examples field={field} />;
       } else {
-        exampleField = <FieldDetail label={label} value={example} />;
+        const label = l('example') + ':';
+        const raw = !!field.in;
+        renderedExamples = <FieldDetail label={label} value={getSerializedValue(field, field.example)} raw={raw} />;
       }
     }
 
@@ -74,10 +76,23 @@ export class FieldDetails extends React.PureComponent<FieldProps, { patternShown
               &gt;{' '}
             </TypeFormat>
           )}
+          {schema.contentEncoding && (
+            <TypeFormat>
+              {' '}&lt;
+              {schema.contentEncoding}
+              &gt;{' '}
+            </TypeFormat>
+          )}
+          {schema.contentMediaType && (
+            <TypeFormat>
+              {' '}&lt;
+              {schema.contentMediaType}
+              &gt;{' '}
+            </TypeFormat>
+          )}
           {schema.title && !hideSchemaTitles && <TypeTitle> ({schema.title}) </TypeTitle>}
           <ConstraintsView constraints={schema.constraints} />
-          {schema.nullable && <NullableLabel> {l('nullable')} </NullableLabel>}
-          {schema.pattern && (
+          {schema.pattern && !hideSchemaPattern && (
             <>
               <PatternLabel>
                 {patternShown || schema.pattern.length < MAX_PATTERN_LENGTH
@@ -100,7 +115,7 @@ export class FieldDetails extends React.PureComponent<FieldProps, { patternShown
         )}
         <FieldDetail raw={rawDefault} label={l('default') + ':'} value={schema.default} />
         {!renderDiscriminatorSwitch && <EnumValues type={schema.type} values={schema.enum} />}{' '}
-        {exampleField}
+        {renderedExamples}
         {<Extensions extensions={{ ...field.extensions, ...schema.extensions }} />}
         <div>
           <Markdown compact={true} source={description} />
@@ -109,7 +124,45 @@ export class FieldDetails extends React.PureComponent<FieldProps, { patternShown
           <ExternalDocumentation externalDocs={schema.externalDocs} compact={true} />
         )}
         {(renderDiscriminatorSwitch && renderDiscriminatorSwitch(this.props)) || null}
+        {field.const && (<FieldDetail label={l('const') + ':'} value={field.const} />) || null}
       </div>
     );
   }
 }
+
+function Examples({ field }: { field: FieldModel }) {
+  if (!field.examples) {
+    return null;
+  }
+
+  return (
+    <>
+      <FieldLabel> {l('examples')}: </FieldLabel>
+      <ExamplesList>
+        {Object.values(field.examples).map((example, idx) => {
+          return (
+            <li key={idx}>
+              <ExampleValue>{getSerializedValue(field, example.value)}</ExampleValue> - {example.summary || example.description}
+            </li>
+          );
+        })}
+      </ExamplesList>
+    </>
+  );
+}
+
+function getSerializedValue(field: FieldModel, example: any) {
+  if (field.in) {
+    // decode for better readability in examples: see https://github.com/Redocly/redoc/issues/1138
+    return decodeURIComponent(serializeParameterValue(field, example));
+  } else {
+    return example;
+  }
+}
+
+
+const ExamplesList = styled.ul`
+  margin-top: 1em;
+  padding-left: 0;
+  list-style-position: inside;
+`;