+++ src/components/Schema/ArraySchema.tsx
@@ -4,7 +4,7 @@ import { Schema, SchemaProps } from './Schema';
 
 import { ArrayClosingLabel, ArrayOpenningLabel } from '../../common-elements';
 import styled from '../../styled-components';
-import {humanizeConstraints} from "../../utils";
+import { humanizeConstraints } from "../../utils";
 import { TypeTitle } from '../../common-elements/fields';
 
 const PaddedSchema = styled.div`
@@ -13,13 +13,18 @@ const PaddedSchema = styled.div`
 
 export class ArraySchema extends React.PureComponent<SchemaProps> {
   render() {
-    const itemsSchema = this.props.schema.items!;
-    const itemConstraintSchema = (
-      min: number | undefined = undefined,
-      max: number | undefined = undefined,
-    ) => ({ type: 'array', minItems: min, maxItems: max });
+    const schema = this.props.schema;
+    const itemsSchema = schema.items;
 
-    const minMaxItems = humanizeConstraints(itemConstraintSchema(itemsSchema.schema.minItems, itemsSchema.schema.maxItems));
+    const minMaxItems = schema.minItems === undefined && schema.maxItems === undefined ?
+      '' :
+      `(${humanizeConstraints(schema)})`;
+
+    if (schema.displayType && !itemsSchema && !minMaxItems.length) {
+      return (<div>
+        <TypeName>{schema.displayType}</TypeName>
+      </div>);
+    }
 
     return (
       <div>