+++ src/services/models/Field.ts
@@ -11,6 +11,8 @@ import { RedocNormalizedOptions } from '../RedocNormalizedOptions';
 import { extractExtensions } from '../../utils/openapi';
 import { OpenAPIParser } from '../OpenAPIParser';
 import { SchemaModel } from './Schema';
+import { ExampleModel } from './Example';
+import { mapValues } from '../../utils/helpers';
 
 const DEFAULT_SERIALIZATION: Record<
   OpenAPIParameterLocation,
@@ -46,12 +48,14 @@ export class FieldModel {
   required: boolean;
   description: string;
   example?: string;
+  examples?: Record<string, ExampleModel>;
   deprecated: boolean;
   in?: OpenAPIParameterLocation;
   kind: string;
   extensions?: Record<string, any>;
   explode: boolean;
   style?: OpenAPIParameterStyle;
+  const?: any;
 
   serializationMime?: string;
 
@@ -81,6 +85,13 @@ export class FieldModel {
       info.description === undefined ? this.schema.description || '' : info.description;
     this.example = info.example || this.schema.example;
 
+    if (info.examples !== undefined) {
+      this.examples = mapValues(
+        info.examples,
+        (example, name) => new ExampleModel(parser, example, name, info.encoding),
+      );
+    }
+
     if (serializationMime) {
       this.serializationMime = serializationMime;
     } else if (info.style) {
@@ -101,6 +112,8 @@ export class FieldModel {
     if (options.showExtensions) {
       this.extensions = extractExtensions(info, options.showExtensions);
     }
+
+    this.const = this.schema?.const || info?.const || '';
   }
 
   @action